requirements.txt
flask
flask-socketio
eventlet
gunicorn
qrcode
pillow
groq
pandas
openpyxl
pymupdf
python-docx